requirements.txt
streamlit>=1.28.0
blocket-api>=0.4.3
tenacity>=8.0.0
pydantic>=2.0.0
mysql-connector-python>=8.0.0
pytest>=7.0.0
